fix(oauth): Get Session object after save and early return on start
After we save the session in Set Expiry the session and its pointer are invalid,
as described in the source code of session.save()
We now get a new session after saving the old one.

Return early if the user is not logged in, it's the case if we visited
this for the first time.