fix: make it sync again and save createdAt int the db
if we save the object in this handler and another route wants to save
the classroom at the same time, the last one wins. So now we make it
sync again and save the createdAt in the db to check for it without
calling to gitlab. Now the first request each day is a little bit slower
but this is acceptable instead of transactions or something eles.